從 yahoo 取得前一個交易日的加權指數的成交量

+++ Core/Model/YahooTechData.cs
@@ -23,7 +23,7 @@ namespace Core.Model
     }
     public class ChartModel
     {
-        public List<int> Timestamp { get; set; }
+        public List<long> Timestamp { get; set; }
         public IndicatorModel Indicators { get; set; }
     }
     public class IndicatorModel
